Majority of Deliverable 3 Complete & UI Overhauled. See Description.
List of attendees registered for an event is now viewable and split into pending and approved attendees. ApproveAllRegistrations is confirmed to be functional. Dialog is set up to cooperate with new Firebase layout. TODO: fix EventRegistrationAttendeeViewHolder so it properly displays attendee information and is clickable (opens dialog). Dialog has already been edited to only show accept/reject buttons when attendee is pending.